TN: action classification update
